dapp: fix library linking on solc 0.7+

+++ src/dapp-tests/pass/libraries.sol
@@ -1,18 +1,18 @@
 import "ds-test/test.sol";
 
 library A {
-    function f(uint x) public returns (uint256) {
-        return x * 2;
+    function f(uint128 x) public returns (uint256) {
+        return uint(x) * 2;
     }
 }
 
 contract B is DSTest {
-    using A for uint256;
-    function test_f(uint x) public {
-        assertEq(x * 2, x.f());
+    using A for uint128;
+    function test_f(uint128 x) public {
+        assertEq(uint(x) * 2, x.f());
     }
 
     function testFail_f() public {
-        assertEq(1, uint(1).f());
+        assertEq(1, uint128(1).f());
     }
 }
